move fixtures to common
