+++ e2e/pre/e2e_test.go
@@ -23,7 +23,7 @@ import (
 
 	natsv1alpha1 "github.com/kyma-project/nats-manager/api/v1alpha1"
 	. "github.com/kyma-project/nats-manager/e2e/common"
-	. "github.com/kyma-project/nats-manager/e2e/fixtures"
+	. "github.com/kyma-project/nats-manager/e2e/common/fixtures"
 )
 
 // Constants for retries.
